+++ massiv/src/Data/Massiv/Array/Numeric.hs
@@ -23,6 +23,7 @@ module Data.Massiv.Array.Numeric
   , (.*)
   , (*.)
   , (.^)
+  , (#>)
   , (|*|)
   , multiplyTransposed
   , identityMatrix
@@ -198,7 +199,9 @@ liftNumericArray2M f a1 a2
 (.^) = powerPointwise
 {-# INLINE (.^) #-}
 
--- | Perform matrix multiplication. Inner dimensions must agree, otherwise `SizeMismatchException`.
+-- | Matrix multiplication
+--
+-- Inner dimensions must agree, otherwise `SizeMismatchException`.
 (|*|) ::
      (Mutable r Ix2 e, Source r' Ix2 e, OuterSlice r Ix2 e, Source (R r) Ix1 e, Num e, MonadThrow m)
   => Array r Ix2 e
@@ -211,6 +214,26 @@ liftNumericArray2M f a1 a2
 "multDoubleTranspose" [~1] forall arr1 arr2 . arr1 |*| transpose arr2 =
     multiplyTransposedFused arr1 (convert arr2)
  #-}
+
+-- | Matrix-vector product
+--
+-- Inner dimensions must agree, otherwise `SizeMismatchException`
+--
+-- @since 0.5.2
+(#>) :: (MonadThrow m, Num e, Source (R r) Ix1 e, Manifest r' Ix1 e, OuterSlice r Ix2 e) =>
+        Array r Ix2 e -- ^ Matrix
+     -> Array r' Ix1 e -- ^ Vector
+     -> m (Array D Ix1 e)
+mm #> v
+  | mCols /= n = throwM $ SizeMismatchException (size mm) (Sz2 n 1)
+  | otherwise = pure $ makeArray (getComp mm <> getComp v) (Sz1 mRows) $ \i ->
+      A.foldlS (+) 0 (A.zipWith (*) (unsafeOuterSlice mm i) v)
+  where
+    Sz2 mRows mCols = size mm
+    Sz1 n = size v
+{-# INLINE (#>) #-}
+
+
 
 multiplyTransposedFused ::
      ( Mutable r Ix2 e
@@ -241,8 +264,7 @@ multArrs arr1 arr2 = multiplyTransposed arr1 arr2'
     arr2' = compute $ transpose arr2
 {-# INLINE multArrs #-}
 
--- | It is quite often that second matrix gets transposed before multiplication (eg. A * A'), but
--- due to layout of data in memory it is more efficient to transpose the second array again.
+-- | Computes the matrix-matrix transposed product (i.e. A * A')
 multiplyTransposed ::
      ( Manifest r Ix2 e
      , OuterSlice r Ix2 e