+++ massiv/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Addition of `lowerTriangular` and `upperTriangular`
 * Relax `identityMatrix` type to return an array of any `Num` type, not just `Int`.
 * Addition of `unsafeMakeLoadArrayAdjusted`
+* Add matrix-vector product (`(#>)`)
 
 # 0.5.1